python-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
loguru==0.7.3
httpx==0.28.1
python-json-logger==2.0.7
python-dotenv==1.1.1
PyYAML==6.0.2

# MCP (Model Context Protocol) dependencies
# Note: structlog and aiohttp are optional but recommended for production
# structlog>=24.1.0
# aiohttp>=3.10.0

# MCP Phase 3 - Tool discovery and execution
jsonschema>=4.0.0

# JobSpy for job scraping integration
python-jobspy==1.1.82
pandas==2.3.2

# Web scraping for Glassdoor job descriptions
playwright==1.49.1

# Queue and scheduler dependencies
# Pin to versions available on PyPI to ensure installation succeeds
rq==1.16.2
rq-scheduler==0.13.1
redis==6.1.1

# Force wheel-friendly versions to avoid compiling from source
grpcio==1.67.1
# Must be >=5.26.1 and <6.0dev for grpcio-status
protobuf==5.29.1
grpcio-status==1.67.1
googleapis-common-protos==1.70.0
google-genai==1.32.0

# LLM Provider clients
ollama==0.5.3
openai==2.14.0
boto3==1.42.20
litellm==1.80.11
langfuse>=3.0.0
# Crew orchestration (Legacy)
# [DEPRECATED] crewai==...

# New AI Workflow Standard (2026)
# Standardizing on LiteLLM + Langfuse + ActivePieces Webhooks
# (LangChain/Langflow/LangGraph packages removed to resolve version conflicts)
# For web search tool
tavily-python==0.7.11

# Database connection for direct access
asyncpg==0.30.0
psycopg[binary]==3.2.9

# Vector database client
chromadb == 1.4.0
sentence-transformers==4.1.0
torch==2.2.0
